# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DepartmentHub - Academic Event & Progress Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events with optional filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: departmentId
          in: query
          schema:
            type: integer
        - name: week
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a new event (Faculty/Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get a single event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Event not found
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update an event (Faculty/Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Event updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event (Faculty/Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Event deleted
...
  /progress:
    get:
      operationId: listProgressLogs
      tags: [progress]
      summary: List progress logs
      parameters:
        - name: departmentId
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
        - name: week
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of progress logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProgressLog"
    post:
      operationId: createProgressLog
      tags: [progress]
      summary: Create a progress log (Faculty/Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgressLogInput"
      responses:
        "201":
          description: Progress log created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressLog"
  /progress/{id}:
    get:
      operationId: getProgressLog
      tags: [progress]
      summary: Get a progress log
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Progress log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressLog"
        "404":
          description: Not found
    patch:
      operationId: updateProgressLog
      tags: [progress]
      summary: Update a progress log
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgressLogUpdate"
      responses:
        "200":
          description: Updated progress log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressLog"
    delete:
      operationId: deleteProgressLog
      tags: [progress]
      summary: Delete a progress log
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users (Admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Departments
...
components:
  schemas:
...
    User:
      type: object
      required: [id, email, firstName, lastName, role, isActive, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [ADMIN, FACULTY, STUDENT, MANAGEMENT]
        departmentId:
          type: ["integer", "null"]
        departmentName:
          type: ["string", "null"]
        isActive:
          type: boolean
        lastLogin:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    UserInput:
      type: object
      required: [email, password, firstName, lastName, role]
      properties:
        email:
          type: string
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [ADMIN, FACULTY, STUDENT, MANAGEMENT]
        departmentId:
          type: ["integer", "null"]
    UserUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [ADMIN, FACULTY, STUDENT, MANAGEMENT]
        departmentId:
          type: ["integer", "null"]
        isActive:
          type: boolean

    # Department schemas
...
    Event:
      type: object
      required: [id, title, category, status, priority, startDate, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
          enum: [ACADEMIC, EXTRACURRICULAR, ADMINISTRATIVE]
        status:
          type: string
          enum: [PLANNED, ONGOING, COMPLETED, CANCELLED]
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        createdById:
          type: integer
        createdByName:
          type: ["string", "null"]
        departmentId:
          type: ["integer", "null"]
        departmentName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    EventInput:
      type: object
      required: [title, category, status, priority, startDate]
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [ACADEMIC, EXTRACURRICULAR, ADMINISTRATIVE]
        status:
          type: string
          enum: [PLANNED, ONGOING, COMPLETED, CANCELLED]
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        location:
          type: string
        departmentId:
          type: ["integer", "null"]
    EventUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [ACADEMIC, EXTRACURRICULAR, ADMINISTRATIVE]
        status:
          type: string
          enum: [PLANNED, ONGOING, COMPLETED, CANCELLED]
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        location:
          type: string
...
    ProgressLog:
      type: object
      required: [id, title, weekNumber, year, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        weekNumber:
          type: integer
        year:
          type: integer
        metrics:
          type: ["string", "null"]
        createdById:
          type: integer
        createdByName:
          type: ["string", "null"]
        departmentId:
          type: ["integer", "null"]
        departmentName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    ProgressLogInput:
      type: object
      required: [title, weekNumber, year]
      properties:
        title:
          type: string
        description:
          type: string
        weekNumber:
          type: integer
        year:
          type: integer
        metrics:
          type: string
        departmentId:
          type: ["integer", "null"]
    ProgressLogUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        metrics:
          type: string
...